src/Mint-Y/gtk-3.0/sass/_common.scss line 26: outline-style: none; (was: dashed)
Mint-Yz + Added solid and coloured outlines on buttons, but none on the others. Because i found annoying
to see an outline in Nemo in some weird cases like when deleting a directory and then seeing a useless
dashed outline on the other directory next to it for no reason, or when moving through selected
directories using the arrow keys and then seeing both the selected background color and the dashed line
on top of it.